requirements.txt
python-dotenv
flask
openai==1.35.15
requests
streamlit
beautifulsoup4
chromadb
sentence-transformers
together
tf-keras
